Update 3
Added global variables to store necessary info.  Added function that generates the ssn label text strings and stores them in an array.  Added function that generates the ssn link and name strings and saves them in an array.  Added function that parses the labels into an id and stores them in an array.